start working on Manchester Baby

+++ src/scala/dedesim/Simulator.scala
@@ -14,7 +14,7 @@ object Simulator {
         msg.wireEvent(sim.currentTime, wire.id, wire.getSignal)
         if (wire.parent.cType == 'wires) {
             val ws = wire.parent.asInstanceOf[Wires]
-            msg.wireEvent(sim.currentTime, ws.id, ws.getSignalAsInt, ws.width)
+            msg.wireEvent(sim.currentTime, ws.id, ws.getSignalAsInt.toLong, ws.width)
         }
     }
